+++ src/mandelparams.h
@@ -39,8 +39,8 @@ struct MandelParameters {
     unsigned int bailout;
 
     unsigned int zoom;
-    unsigned int xcoord;
-    unsigned int ycoord;
+    double xcoord;
+    double ycoord;
 
     std::string image_base;
 
@@ -52,7 +52,7 @@ struct MandelParameters {
     MandelParameters(unsigned int xrange, double xl, double xh,
                      unsigned int yrange, double yl, double yh,
                      unsigned int bailout, unsigned int zoom,
-                     unsigned int xcoord, unsigned int ycoord,
+                     double xcoord, double ycoord,
                      std::string image_base, unsigned int cores,
                      constants::COL_ALGO col_algo)
         : xrange(xrange),
