+++ src/mandelzoom.h
@@ -40,8 +40,8 @@ public:
      * @param ycoord Y coordinate in the complex plane
      */
     void calcalute_zoom_cpane(double &xh, double &xl, double &yh, double &yl,
-                              unsigned int zoom, unsigned int xcoord,
-                              unsigned int ycoord, unsigned int width,
+                              unsigned int zoom, double xcoord,
+                              double ycoord, unsigned int width,
                               unsigned int height);
 
 private:
