+++ src/test/buffwriter_mock.cpp
@@ -28,7 +28,7 @@ void BuffwriterMock::write_buffer(){};
 std::string BuffwriterMock::test_filename_patterns(
     std::string filename_pattern, unsigned int bailout, unsigned int xrange,
     unsigned int yrange, unsigned int zoom, unsigned int cores,
-    unsigned int xcoord, unsigned int ycoord, double z_real_min,
+    double xcoord, double ycoord, double z_real_min,
     double z_real_max, double z_ima_min, double z_ima_max)
 {
     return this->out_file_name(filename_pattern, bailout, xrange, yrange, zoom,
