+++ src/test/buffwriter_mock.h
@@ -31,8 +31,8 @@ public:
     std::string test_filename_patterns(std::string filename_pattern,
                                        unsigned int bailout, unsigned int xrange,
                                        unsigned int yrange, unsigned int zoom,
-                                       unsigned int cores, unsigned int xcoord,
-                                       unsigned int ycoord, double z_real_min,
+                                       unsigned int cores, double xcoord,
+                                       double ycoord, double z_real_min,
                                        double z_real_max, double z_ima_min,
                                        double z_ima_max);
